+++ Ecommerce_Project/Models/Prices/ProductPrice.cs
@@ -3,8 +3,8 @@
     public class ProductPrice
     {
         public int Id { get; set; }
-        public double MinimumPieces { get; set; }
-        public double Price { get; set; }
+        public long MinimumPieces { get; set; }
+        public decimal Price { get; set; }
 
         // Foreign keys
         public Product BodyProduct { get; set; }
